Add additional test logic and reorganize some files
