+++ tests/helpers/helper_classes/RequestValidator.py
@@ -65,9 +65,7 @@ from endpoints.schema_config.instantiations.data_requests.related_locations.dele
 from endpoints.schema_config.instantiations.data_requests.related_locations.post import (
     DataRequestsRelatedLocationsPostEndpointSchemaConfig,
 )
-from endpoints.schema_config.instantiations.data_sources.by_id.get import (
-    DataSourcesByIDGetEndpointSchemaConfig,
-)
+from endpoints.instantiations.data_sources_.get.by_id.schema_config import DataSourcesByIDGetEndpointSchemaConfig
 from endpoints.schema_config.instantiations.data_sources.by_id.reject import (
     DataSourcesByIDRejectEndpointSchemaConfig,
 )
