+++ tests/integration/data_sources/put/test_approval_status.py
@@ -1,7 +1,5 @@
 from db.enums import ApprovalStatus
-from endpoints.schema_config.instantiations.data_sources.by_id.get import (
-    DataSourcesByIDGetEndpointSchemaConfig,
-)
+from endpoints.instantiations.data_sources_.get.by_id.schema_config import DataSourcesByIDGetEndpointSchemaConfig
 from tests.helpers.asserts import assert_contains_key_value_pairs
 from tests.helpers.constants import DATA_SOURCES_BASE_ENDPOINT
 from tests.helpers.helper_classes.test_data_creator.flask import (
