+++ tests/integration/data_sources/put/test_basic.py
@@ -8,9 +8,7 @@ from db.enums import (
     RetentionSchedule,
     URLStatus,
 )
-from endpoints.schema_config.instantiations.data_sources.by_id.get import (
-    DataSourcesByIDGetEndpointSchemaConfig,
-)
+from endpoints.instantiations.data_sources_.get.by_id.schema_config import DataSourcesByIDGetEndpointSchemaConfig
 from middleware.enums import RecordTypes
 from tests.helpers.asserts import assert_contains_key_value_pairs
 from tests.helpers.common_test_data import get_test_name
