+++ tests/middleware/access_logic/get_authentication/constants.py
@@ -1,0 +1,1 @@
+PATCH_ROOT = "middleware.security.auth.helpers.get_header_auth_info"
