+++ tests/middleware/access_logic/mocks/get_access_info_from_jwt_or_api_key.py
@@ -1,0 +1,10 @@
+from unittest.mock import MagicMock
+
+from tests.helpers.DynamicMagicMock import DynamicMagicMock
+
+
+class GetAccessInfoFromJWTOrAPIKeyMocks(DynamicMagicMock):
+    get_user_email_from_api_key: MagicMock
+    get_jwt_identity: MagicMock
+    AccessInfo: MagicMock
+    get_user_permissions: MagicMock
