+++ tests/middleware/access_logic/mocks/get_api_key_from_request_header.py
@@ -1,0 +1,8 @@
+from unittest.mock import MagicMock
+
+from tests.helpers.DynamicMagicMock import DynamicMagicMock
+
+
+class GetAPIKeyFromRequestHeaderMock(DynamicMagicMock):
+    get_authorization_header_from_request: MagicMock
+    get_api_key_from_authorization_header: MagicMock
